Embed state mod in database mod
The `Swap` struct is now database specific, no need to have a 20 lines
 module.
