+++ swap/src/database.rs
@@ -1,12 +1,10 @@
 use anyhow::{anyhow, bail, Context, Result};
-use serde::{de::DeserializeOwned, Serialize};
-use state::Swap;
-use std::path::Path;
+use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use std::{fmt::Display, path::Path};
 use uuid::Uuid;
 
 mod alice;
 mod bob;
-pub mod state;
 
 pub use alice::*;
 pub use bob::*;
@@ -85,6 +83,33 @@ where
     T: DeserializeOwned,
 {
     Ok(serde_cbor::from_slice(&v)?)
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+pub enum Swap {
+    Alice(Alice),
+    Bob(Bob),
+}
+
+impl From<Alice> for Swap {
+    fn from(from: Alice) -> Self {
+        Swap::Alice(from)
+    }
+}
+
+impl From<Bob> for Swap {
+    fn from(from: Bob) -> Self {
+        Swap::Bob(from)
+    }
+}
+
+impl Display for Swap {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Swap::Alice(alice) => Display::fmt(alice, f),
+            Swap::Bob(bob) => Display::fmt(bob, f),
+        }
+    }
 }
 
 #[cfg(test)]
