--- swap/src/database/state.rs
@@ -1,30 +1,0 @@
-use crate::database::{Alice, Bob};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
-
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
-pub enum Swap {
-    Alice(Alice),
-    Bob(Bob),
-}
-
-impl From<Alice> for Swap {
-    fn from(from: Alice) -> Self {
-        Swap::Alice(from)
-    }
-}
-
-impl From<Bob> for Swap {
-    fn from(from: Bob) -> Self {
-        Swap::Bob(from)
-    }
-}
-
-impl Display for Swap {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Swap::Alice(alice) => Display::fmt(alice, f),
-            Swap::Bob(bob) => Display::fmt(bob, f),
-        }
-    }
-}
